ActorSalary button finished

+++ MoviesData/IActorRepository.cs
@@ -6,7 +6,7 @@ namespace MoviesData
 {
     public interface IActorRepository
     {
-        float ActorTotalSalary(string firstName, string lastName);
+        double ActorTotalSalary(string firstName, string lastName);
         IReadOnlyList<(Actor, Actor, Movie)> ActorInCommon(string firstName, string lastName);
     }
 }
